Add more log messages to worker.py (#483)
It can sometimes be helpful to determine what version of Sematic a
particular pipeline was running with when looking in the logs. This PR
adds that info. Also, to diagnose whether there were new pods created,
or help with k8s debugging, it can be helpful to view the pod name. This
PR also adds that to the logs.

Testing
-------
Executed testing pipeline. View logs
[here](https://dev1.dev-usw2-sematic0.sematic.cloud/pipelines/sematic.examples.testing_pipeline.pipeline.testing_pipeline/982287fea62d49f09670e7673c4524e3#run=027010013add4641a0accecf80ecd3c9&tab=logs)