Enable dataclass compatibility (#484)
Sometimes dataclasses change over time, and it would be helpful if
values deserialized using an old version of a class could be
deserialized using a new version of the class. One would think that if
you add a new field to a dataclass, it should be able to deserialize
something written when the class did *not* have that field--provided
that the new class has a default value for the field. This is how
protobuf compatibility works, for example. However, our code did not
support this use case until this PR.

One place where these kind of incongruences can show up is between
clients on an old version of Sematic and servers on a new version of
Sematic. Clients and servers share some dataclasses, like resource
requirements. This makes it safe to add new fields to dataclasses,
provided that you specify a default for the new fields.