Make _getitem importable on graph cloning (#499)
Fixes #494.

When re-running a pipeline that included `_getitem`, during the cloning
process the func would be imported. As `_getitem` is declared inside
another plain Python function, it is not available for import.

This does not reproduce for `_make_list` or for `_make_tuple` because
their enclosing functions have the same respective names. `_getitem` is
declared inside a function called `__getitem__`, which is then set as
the homonymous utility function on the `Future` class.

This PR introduces a `_getitem` function which wraps `__getitem__`,
which can be imported in its stead.

The testing pipeline is also updated to provide the ability to
explicitly include the `_make_list`, `_make_tuple`, and `_getitem`
virtual funcs.

Also, `_getitem` is added to the list of "hidden" funcs in the
Dashboard, alongside the other two. This had been an oversight.

Reproduced locally using the testing pipeline enhancement, and validated
locally and in the cloud that the re-run feature works with this fix:

```bash
$ bazel run sematic/examples/testing_pipeline -- --virtual-funcs
$ bazel run sematic/examples/testing_pipeline -- --virtual-funcs --rerun-from=XXX
```


![image](https://user-images.githubusercontent.com/1894533/214453598-718b6b4c-3f4d-4343-b057-d00fe5fac95b.png)