Support for non-deterministic cached builds in Bazel (#500)
Resolves #418, by changing up how we pass in the Docker container image
uris to the launch script in Bazel.

Originally, we were writing a file with the name
`_{tag}_push_at_build.uri` during the Bazel build stage, and this file
got read by the launch script in `container_images.py` so that the
launch script can tell the server the tags and images for the run.

Since this was able to only be written during the build stage, it did
not work for the case where the image is not actually the same but we
want to rerun using the previously pushed image uri. So now we instead
rely on a digest file that is written by Bazel itself during the
container push step. The contents of this file are read into an encoded
string stored in the `SEMATIC_CONTAINER_IMAGE_URIS` env var, which is
then read by the launch script and sent to the server instead.

TODOs:
- [x] test with multiple container images in the same run, need to find
out how to test this
- [x] verify that this does indeed fix the reporting user's issue by
reproducing it somehow (can we even do that?)

Co-authored-by: augray <augray@users.noreply.github.com>